Refactoring image-package
