+++ src/main/java/com/ilnitsk/animusic/album/dto/AlbumItemDto.java
@@ -1,6 +1,6 @@
 package com.ilnitsk.animusic.album.dto;
 
-import com.ilnitsk.animusic.image.CoverArtDto;
+import com.ilnitsk.animusic.image.dto.CoverArtDto;
 import lombok.Data;
 
 @Data
