+++ src/main/java/com/ilnitsk/animusic/playlist/dao/Playlist.java
@@ -1,7 +1,7 @@
 package com.ilnitsk.animusic.playlist.dao;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.ilnitsk.animusic.image.CoverArt;
+import com.ilnitsk.animusic.image.dao.CoverArt;
 import com.ilnitsk.animusic.user.dao.User;
 import jakarta.persistence.*;
 import lombok.*;
