+++ src/main/java/com/ilnitsk/animusic/playlist/dto/PlaylistDto.java
@@ -1,6 +1,6 @@
 package com.ilnitsk.animusic.playlist.dto;
 
-import com.ilnitsk.animusic.image.CoverArtDto;
+import com.ilnitsk.animusic.image.dto.CoverArtDto;
 import lombok.Data;
 
 import java.util.List;
