+++ src/test/java/com/ilnitsk/animusic/services/AnimeServiceTest.java
@@ -5,7 +5,7 @@ import com.ilnitsk.animusic.anime.repository.AnimeRepository;
 import com.ilnitsk.animusic.anime.service.AnimeService;
 import com.ilnitsk.animusic.exception.AnimeNotFoundException;
 import com.ilnitsk.animusic.exception.BadRequestException;
-import com.ilnitsk.animusic.image.ImageService;
+import com.ilnitsk.animusic.image.service.ImageService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
